fix: rename package to todo for more readability
