+++ app/src/test/java/com/codesoom/todo/AppTest.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment;
+package com.codesoom.todo;
 
 import org.junit.jupiter.api.Test;
